Annotate API for LE Scanner and LE Advertiser getters with @Nullable
This explicitly indicates in the API that these methods can return null
if Bluetooth is not `ON` or `BLE_ON` as seen in the implementation of
`BluetoothAdapter#getLeAccess`.

This annotation specifically will allow clients to gracefully handle
instead of leaving an ambiguity on the nullness of `bluetoothLeScanner`
and `bluetoothLeAdvertiser`.

FCRS_CODE: mhw0wj48y2yak8

Bug: 394958283
Test: atest BluetoothJavaUnitTests
Flag: EXEMPT PURE_REFACTOR
Change-Id: I16e7e647b74fc048295a5e6b35c0b3b8f0db0ebb