Disconnecting doesn't wait for state change from native
Problem: Transitions to the Disconnecting state always passed straight
through to the Disconnected state. This puts our state machine in java
out of sync with the state being communicated from native, contributing
to odd lifecycle bugs. This is particular problematic, as we remove the
state machine when entering disconnected. This bug makes it so we remove
the state machine too early, then get the DISCONNECTED state from
native, then re-create the state machine, just to re-destroy it.

Solution: Don't immediately transition to Disconnected. Wait for the
Disconnected state change from native. Bring the Disconnecting State
implementation up to date with other in the stack, using a disconnecting
timeout. Add tests to prove stability.

Flag: EXEMPT BUGFIX
Bug: 460427587
Test: atest com.android.bluetooth.a2dpsink
Change-Id: Ia3f3e14b2bc00a76df52631727f695a94016c641