Fix null pointer dereferencing in btm_pm_proc_mode_change fn
The current PM code in the stack incorrectly assumes that if the
registration mask bit is set, the associated callback (pm_reg_db.cback)
is guaranteed to be valid.

This assumption leads to a SIGSEGV crash when a race condition or
asynchronous event (e.g., during Bluetooth factory reset) prematurely
sets pm_reg_db.cback to null while a mode change event is being
processed on the main thread.

This change adds a proper null check to the callback before invoking the
function. The same check is applied to two other vulnerable locations
where pm_reg_db.cback is accessed.

Bug: 459820263
Test: mmm packages/modules/Bluetooth
Flag: EXEMPT BUGFIX
Change-Id: Ib2db7063fc64fea220b6b6ace4579f5e394a6c25