Flagged Change: Mark port not in use on release
When we release a port, we currently do not change the in_use flag to
false. This means that server ports where the port handle and some
default values are kept are still labeled as "in use"

If the DLCI happens to match the DLCI of a new server port, and this
port has a port_handle that's less than that of the new port, we can run
into a crash. This is an attempt to prevent the crash, with logs to help
confirm if the fix works.

Bug: 391740410
Bug: 460549793
Test: m com.android.bt
Flag: com.android.bluetooth.flags.mark_port_not_in_use_on_release
Change-Id: I4969f28bb81ac7bf4cacb130bfeaca1027587070