SystemServer own jarjaring
It seems we must not share some kotlin and androidx dependencies into
system server class path.
The option is to jarjar those. But, since Bluetooth jarjaring is shared
between the apk-in-apex, the framework sdk and the system server, doing
a full jarjar of androidx + kotlin + kotlinx is bringing close to 2MB of
increased in size.
And it still doesn't boot.

Instead, we could be doing a different jarjaring inside the system
server, that only has the concerned classes, and only here we would
jarjar more aggressively.

Bug: 457805965
Test: m .
Flag: NONE updating jarjar classes for conformance issues
Change-Id: I6a6a6964a215578f35337b611b2d5f44bc045b0c