Rename ScanManager scan timeout method for consistency
The name `regularScanTimeout` was confusing. It doesn't reflect that it
is a method called when the timeout _occurs_. The format is now aligned
with all the other methods under `ClientHandler` and on
`Flags.scanControllerThread` cleanup, can possibly be renamed to be more
accurate alongside all the other `handle...` methods.

Bug: 394958283
Test: atest ScanManagerTest
Flag: EXEMPT PURE_REFACTOR
Change-Id: I4e3f394f723ba719587ffc0c58c43c2af804ea3c