Introduce ScanUtilTest
Even though I am not a big fan of testing string outputs in unit tests,
the outputs of these Framework API constants are unlikely to change.
And so, to fix the annoying "Warning: Missing high value tests" in
`ScanUtil` CLs, this is just basic tests to increase the test coverage
of `ScanUtil.kt`.

Bug: 429793161
Test: atest ScanUtilTest
Flag: EXEMPT TEST_ONLY
Change-Id: I2d8139c8f2d934acb1cd8fbd19d65dc12c13e0e8