leaudio: Fix metadata update before stream starts.
If stream configuration happens in two steps e.g. first it is
configured in the preparation step which can take ASEs either
CODEC_CONFIGURED or QOS_CONFIGURED state and then it is Enabled when
Audio HAL resumes the stream, it might happen that Metadata changes but
stream configuration stays.

In this case, stack might send Enable CTP with not valid metadata which
might cause IOP issues.

This CL fixes that.

Bug: 448514133
Flag: EXEMPT BUGFIX
Test: atest bluetooth_le_audio_test

Change-Id: I94a3449b8ec5e362f9056ec584f242e22bfdbcad