Remove TODO for calling `enforceCallingUid` from Utils.java
I've gone over the whole original bug (b/188391719) and it looks like
this was added as a _temporary_ suspension (ag/15260674) after a 2nd
revert due to too many unverified source callers of public APIs.

However, afterwards, a solution was added in frameworks/base ag/15275356
where the call to `enforceCallingUid()` was added directly to the parcel
constructor of `AttributionSource.java`.

It is my understanding that hence *(as it has not been done so in the
last 5 years+)*, `enforceCallingUid` will not be needed to be called
within `Utils.java` and so we can cleanup the TODO as well.

Bug: 188391719
Test: atest BluetoothJavaUnitTests
Flag: EXEMPT PURE_REFACTOR
Change-Id: I9ab0128af8a6ac9a7d14e3ec3424ab0300265c26