hid host: only reconnect ALLOWED devices when loading bonded devices
We always trigger the background connection when loading the list of
bonded HoGP devices, regardless of whether connection policy allows it
or not. This would cause the peer to be able to reconnect on the
baseband layer briefly before we reject the connection at the profile
level, and that might cause the BTA HH state machine to desync.

This CL prevents that by not automatically triggering background conn
for each new device. Instead, the caller needs to specifically call
BTA_HHOpen to put the device into the accept list while also updating
the state machine.

Bug: 458253534
Bug: 459303232
Test: m com.google.android.bt
Test: Manually, verify HID device is not put into accept list when
      policy is set to forbidden and BT is powered off -> on
Flag: com.android.bluetooth.flags.hogp_cancel_gatt_if_policy_forbidden
Change-Id: I8a7a20b172917ead8642169d466d930955e0e77a