Allow all `ScanBinder` calls directly if already on Scan thread
`BassClientService` has an [deprecated] way of using:
`PeriodicAdvertisingManager mPeriodicAdvertisingManager`. In that case,
the calls go to framework (on scan thread) and come back to `ScanBinder`
still on the scan thread. And since both are within the same process,
there is no IPC hop and a binder thread hop does not happen.
This is a temporary solution until the flag
leaudioBroadcastImproveSourceOperations is removed and
`BassClientService` only uses internal `ScanController`.

Fix: 461328929
Fix: 461400415
Fix: 460606932
Bug: 397863857
Bug: 444010402
Test: atest com.android.bluetooth.le_scan
Flag: com.android.bluetooth.flags.scan_controller_thread
Change-Id: Ia8f6e269c082d4d31893bd7ca35194a7c369a42f