Remove BluetoothAdapter.getDefaultAdapter from ScanFilter.Builder
This will ensure ScanFilter.Builder to be a POJO class.

The responsibility of checking for offloaded Transport Discovery Data
scan support is removed from the framework and given to the developer.
The developer can utilize the existing API
  `BluetoothAdapter#getOffloadedTransportDiscoveryDataScanSupported()`
to confirm the support before passing it over in the builder.

FCRS_CODE: mmbswr13hiboc6

Bug: 420601242
Test: atest ScanBinderTest
Test: atest com.android.bluetooth.le_scan
Flag: EXEMPT PURE_REFACTOR
Change-Id: I352a5fb10107490e59a6bdaf6ac2be3d0f7d0854