ScanController uses AdapterService#getRemoteDevice instead of framework
While `AdapterService` itself does use the framework (:)) as the
constructor of `BluetoothDevice` is not public, this at least allows
`ScanController` to optimize the usage of a cached `BluetoothDevice`
object in AdapterService#getRemoteDevice if available in
`mRemoteDevices` when the flag `retainAddressType` is true.

This also allows the removal of depending on `BluetoothManager` and
`BluetoothAdapter` with `ScanController`.

Bug: 377424060
Test: atest BluetoothJavaUnitTests
Flag: EXEMPT PURE_REFACTOR
Change-Id: Ie522f03747f781d5a0e7e1732c4119699bbde504