Fix crash due to processing command response post shutdown
After stack handler shutdown, if any command response is received that
can be processed but when any action will be posted on the handler it
will lead to the crash. Previously handler was not nullptr after
~impl(), so the post call from ContextualCallback was a success, but the
internal mutex was destroyed and hence the crash was during accessing
that. Marked the handler as nullptr now, so we will see null pointer
dereference with
`(*command_queue_.front().GetCallback<CommandCompleteView>())(--)` call
now. Also, the handler shut down was not captured during this processing
because `command_queue_` was only checked as empty() which doesnt
guarantee that HciLayer is still alive.

Bug: 459958923
Bug: 461413549
Test: m com.google.android.bt
Flag: EXEMPT BUGFIX
Change-Id: I8c4d121a3aa695cc7f284249b3cf680280e93fbe