Minimize usage of `getDefaultAdapter()` by injecting BluetoothAdapter
The main usage of this comes from within our app via
`AdapterService.getRemoteDevice` and its overloaded. While internally,
we should be passing around the `BluetoothDevice` object instead of
recreating it, unfortunately that is a massive work for another time.

While this makes the `BluetoothDevice` less of a POJO... this will
substantially lowers the number of
> `"BluetoothAdapter.getDefaultAdapter is deprecated.", new Throwable()`
from being logged.

Further improvements following ag/37059287.

Bug: 420601242
Test: atest BluetoothJavaUnitTests
Flag: EXEMPT PURE_REFACTOR
Change-Id: I147d63d2ccb4fdc5051063fb5fb1e0ad79fe30c2