Validate device configuration at boot
Bluetooth apk name is different when packaged as a mainline module.
Some components of the system (permission, telephony) will rely on a
global resource to know the name of the Bluetooth package and grant some
permissions.

Instead of having a late check on the Bluetooth state because it doesn't
work. Let's have a boot time check that will prevent boot with an
explicit error message.

Bug: 467963541
Bug: 460289219
Test: atest ServiceBluetoothRoboTests
Flag: com.android.bluetooth.flags.validate_bluetooth_name_in_platform_config
Change-Id: I6a6a69646e8f2ac74e4890ee537a8a0b73b86a48