Storage: Store unconfigured sink audio policy
The previous database was storing an UNCONFIGURED audio policy for each
devices. During the initial migration to the new datastore, the default
of UNCONFIGURED was silently migrated as NOT_ALLOWED.

This is not desired.

There is no strict guarantee that UNCONFIGURED need to be stored, but in
case we need it, we should probably have the database ready

----

About the type change in the storage:
This CL is not Wire safe, but it is Wire compatible. And since we are
moving from bool to enum, this is fine.

Here is the mapping of the data:
false -> 0 -> UNCONFIGURED
true  -> 1 -> ALLOWED

Aka, we are preserving any allowed setting, and we are deleting any
not allowed setting on purpose. Indeed, the migration from the database
over populated this field with false.

The side effect is that we no longer need the field to be optional (wire
safe change) because UNCONFIGURED is not send on the wire anyway

Bug: 461053815
Fix: 461053815
Test: atest BluetoothJavaUnitTest
Flag: com.android.bluetooth.flags.mainline_beta_storage
Change-Id: I6a6a69643cbae3b3cc29be0f63659dc3de291e81