broadcaster: Confirm stream request after data path setup
Audio HAL was failing to confirm the streaming establishment following a
'Resume request' because the Data Path establishment was not complete.
Currently, the signalling establish callback (BIG & Data Path) is
triggered, but the data path being incomplete prevents the confirmation
callback from executing.
This leaves the Audio HAL in a no-response state, causing an audible gap
in sound.

This commit updates the logic to ensure the streaming confirmation is
only dependent on the necessary state/callbacks and does not stall due
to an incomplete Data Path establishment when streaming *should* proceed,
thereby preventing the timeout and restoring continuous audio.

Bug: 462041567
Bug: 465187086
Bug: 462672161
Test: atest bluetooth_test_broadcaster
Test: atest bluetooth_test_broadcaster_state_machine
Flag: com.android.bluetooth.flags.leaudio_fix_stream_confirm_datapath_race
Change-Id: I92d9e7d17be9045d0c4bda28877cbc95f14b9013