Fix tests for IRK reset
Failed flag advancement for btif_storage_cycle_irk turns out to be due
to test failures introduced during an earlier revision of the CL but
masked by the flag.  We need to move the test to btif_dm_test.cc, which has a valid GD stack, and adjust the condition checks.

(the test exercises code in both files so I believe this is a valid
location for it)

Test: atest BtifDmWithStackTest#test_btif_dm_reset_irk
Bug: 322326244
Bug: 372714979
Flag: com.android.bluetooth.flags.btsec_cycle_irks
Tag: #security
Change-Id: I2892fbb5b45ed89f8365ae52380a9baa3eae879f