Delete flag dump from Bluetooth
You want to use `adb shell aflags list` instead.

We used to not be able to trust platform flag value.
We used to have to do reflection to manually call the flag and get it's
value.

Both of these issues are fixed for a while, but we kept our manual dump.

But now, with the proper jarjaring and proper compression, we no longer
knows about the list of bluetooth flags, and the dump is showing
nothing.
We could still print some flags, but it would be restricted to the flags
that the system server is using, so the purpose is very low

Bug: 311772251
Test: m .
Test: manually looking at the dumpsys
Flag: EXEMPT BUGFIX
Change-Id: I6a6a6964b880ef21e4c036f7e950a05e7a39eea8