Fix NPE in RemoteDevicesTest by stubbing getRemoteClass
The _real_ fix would be to make sure that `RemoteDevices.java` should
not leak listening to `WatchConnectionStateListener` and causing the
function to be called after test is finished and mocks are cleaned up;
But that's a _harder_ issue to fix. This quick fix should hopefully at
least resolve the 1.3% failure rate of the test.

Fix: 463069355
Test: atest --iterations 20 BluetoothJavaUnitTests:RemoteDevicesTest
Flag: EXEMPT TEST_ONLY
Change-Id: I20e6ca948160dff7d1ad3e8305bfad1ea11129aa