Enabled HFP voice recognition configuration
Tests were using dead code setForceScoAudio(true) which passed the isScoAcceptable check in HeadsetService. setForceScoAudio() was removed and tests to be updated to fulfill one of the prereqs, either being in a call or in voice recognition for SCO to be started up.

so enabling the HFP voice recognition configuration for test purpose.

Bug: 462341285
Test: atest -v HfpTest
Flag: TEST_ONLY
Change-Id: I43bc1a8bf30423f522a9c30506629cd9f33ee921