Use identity address for ScanFilter if it has identity address
For a bonded device, we always need to pass the identity address to the
APCF filter to scan the device. Therefore adding a code snippet to
convert a pseudo address to the identity address if given. If no
matching identity address is found, the added call will have no effect.

Bug: 194260436
Bug: 486698443
Test: m -j
Flag: com.android.bluetooth.flags.convert_pseudo_address_to_identity
Change-Id: If3acc6082871cb9541274a0444428c18f78a9ca7