Fix out of memory crash in cover art.
Some cars require a cover art with a smaller size than the default
200*200 pixel size. A change made it so that we don't check for pixel
size in descriptors anymore which is an issue as some cars will send
requests for huge cover arts (65535*65535) that lead to an out of
memory crash on the phone side. This change makes the flag still
check that the descriptor pixel count is lower than the default if
not equal.

Bug: 454961521
Flag: EXEMPT BUGFIX
Test: m .
Change-Id: I4074f4d296bef4092cfefe24e0335ac9e904688b