leaudio: Fix reporting Disconnected state from Native
In general, Native shall report DISCONNECT state to Java when
1) Device was previously CONNECTED
2) User request connection and this connection failed. There are two possible failures here
a) The ACL connection is not happening.
b) The ACL connection succeed but it is disconnected when device does additional operations in
Without this patch it could happen that Native would not report DISCONNECT state to Java while being in CONNECTED_BY_USER_GETTING_READY state.

Note: If Java is notified about DISCONNECT state in other circumstances, that event is ignored.


Below bug is about DISCONNECT event being not reported as in case 2a).
This CL fixes this.

With this fix, native will also not send unnecessary DISCONNECT Events which previously were ignored by Java anyways e.g.
when

status == GATT_ILLEGAL_PARAMETER &&
          (leAudioDevice->GetConnectionState() == DeviceConnectState::CONNECTING_AUTOCONNECT)


Bug: 408227755
Test: atest bluetooth_le_audio_client_test
Flag: EXEMPT BUGFIX
Change-Id: Ia550baa15bf452cf1628cc22ab79cad071483253