Broadcaster: Add Audio Active State to BASE
According to the Public Broadcast Profile (PBP) v1.0.1 specification,
the Audio_Active_State LTV can be placed in the metadata of a Public
Broadcast Announcement and/or a Basic Audio Announcement (BASE).

When in a Public Broadcast Announcement, it applies to all BISes in the
BIG. When in the BASE structure, it applies to all BISes within a
specific subgroup.

This change adds support for including the Audio_Active_State in the
BASE metadata for each subgroup.

Bug: 463606876
Bug: 463606623
Flag: com.android.bluetooth.flags.leaudio_broadcast_extend_audio_active_state
Test: atest bluetooth_test_broadcaster --host
Change-Id: Ic365a2e5e9d240d9158722f6b595a11f7149a27d