Storage: delay initialization until offToBleOn
Despite initializing the Storage as soon as we have a context is the
correct thing to do, we are in a weird position now since some devices
try to re-use the process end re-create an adapter service on the same
process. Without properly turning off the previous process.

This is not correct.

Unfortunately, there is too much things to fix and those are out of
Bluetooth control if we want to do things properly. Let's first aim to
reduce the noise generated and look into a correct fix later

Flag: com.android.bluetooth.flags.mainline_beta_storage
Bug: 463771626
Bug: 430166215
Test: atest BluetoothJavaUnitTests
Change-Id: I6a6a6964a6516b3596a06c8c32be562f094d42d2