audio_hal_interface: Ignore unknown codecs in offloading preferences
The codec types that are received in the codec offloaded capabilities
are provided by the audio framework by the API
AudioManager.getHwOffloadFormatsSupportedForA2dp and are configured
by the vendor. Thus it is unlikely than an invalid value will be provided.

Rather, unknown values correspond to exclusive codecs that were added by
the vendor through extensive stack modifications.

Supporting these codecs is not possible with the legacy offloading HAL,
instead the codec extensibility APIs were developed for this purpose.

Bug: 441191785
Test: Manual
Test: m com.android.bt
Flag: EXEMPT BUGFIX
Change-Id: I1993fa12d33d0f83ef869dd4662163225a25fb74