Fix GattServerManager death recipient registration
Pass the ActionOnDeathRecipient instance directly to linkToDeath.
Previously, this was passed as a lambda, which resulted in SAM
conversion treating the constructor call as the `binderDied()`
implementation. Consequently, the cleanup logic was never actually
registered or executed.

I've also confirmed that the new "low value" unit test is actually
pretty _valuable_. It fails without the fix in `GattServerManager` prior
to this change and passes with this fix.

Bug: 462054316
Bug: 454373221
Bug: 460158187
Test: atest GattServerManagerTest
Flag: EXEMPT BUGFIX
Change-Id: If19e8db6a848d600155bfa3dc6df5d3c97372a82