Allow for higher Bluetooth advertising Tx power.
The old TX_POWER_MAX constant is deprecated since it is set to 1 (dBm)
but is certainly in use in many apps. Instead a new constant
TX_POWER_MAX_AVAILABLE is added to request the maximum the system can
support.

The Bluetooth specification limits the maximum to 20 dBm. In this CL we
are still limiting to 1 dBm, but subsequent CLs will add a system
property to limit this on certain platforms (small battery devices for
example) as well as capping to the level allowed in the region (10 dbM
for most regions).

Bug: 464003245
Bug: 462097899
Flag: com.android.bluetooth.flags.max_advertising_power
Test: atest BluetoothJavaUnitTests
FCRS_CODE: migmlfo1r7o9di

Change-Id: I969d2d46427496c2b2e45413b3379efd68954261