le_audio: use highest sample rate for mic
Currently the sink sample rate is hard coded to the value corresponding
to the conversational context. This value can be smaller than the actual
sample rates when resolved to the live context, and due to only
upsampling being supported, this breaks the decoder.

In this CL, we choose the highest possible sample rate.

Bug: 442702393
Bug: 464457342
Test: m -j com.google.android.bt
Test: Record LEA SW with Samsung Buds 3 Pro
Flag: com.android.bluetooth.flags.le_audio_use_highest_sample_rate_for_mic
Change-Id: I9daa1a2541e71086b1767f13da5b8e8fd941ee7f