Replace base::Callback with modern alternatives (24,25)
base::Callback is deprecated and should be replaced.
Make the address_callbacks_ map with callbacks in
BleAdvertiserInterfaceImpl store OnceCallbacks.
The callbacks are removed immediately after the first usage,
so it should be safe.

jni_thread_wrapper is a helper method for running callbacks
on the jni thread.
The do_in_jni_thread method used internally already works only
with OnceCallbacks, so it makes sense to propagate that restriction
to jni_thread_wrapper.

Bug: 303363630
Test: atest packages/modules/Bluetooth
Flag: EXEMPT PURE_REFACTOR

Change-Id: I5503ff78f73bb32464302633b1e4222ec23badd8