hh: propagate handshake callback for set_idle and get_idle
The hid device replies set_idle and get_idle with handshake, like they
do with set_report and get_report. However we only propagate the
handshake to the java layer on the latter.

This CL makes it so we do the same for get_idle and set_idle.

Bug: 464954600
Bug: 464973851
Test: m com.google.android.bt
Flag: com.android.bluetooth.flags.hid_propagate_idle_handshake
Change-Id: I1b20cfcd3eb6f98fe323f845cd845cc68b548d0f