leaudio: Improve CIG removal behavior
With this CL we make sure that cig.cises lifetime ends when CIG is
removed.
Before this CL stack could clear cig.cises in a believe that CIG Remove
will always succeed in the Controller which is not true. That could lead
to invalid state and later could end with an assert.
This fixes behavior when controller for some reason will not allow to
remove CIG. Code will try to reuse it or will try to remove it again.

Bug: 457842806
Bug: 465088801
Flag: com.android.bluetooth.flags.leaudio_fix_clear_cises_in_the_cig
Test: atest bluetooth_le_audio_test
Change-Id: I3d36ff8d60d8612e71d42cb81436847376d98a7f