ActiveDeviceManager: change hasFallbackDevice to stopAudio for LEA
Clearly state the expected outcome and function of parameter
for better readability.
Previously hasFallbackDevice indicated to audio framework to NOT
suppress noisy intent and therefore to NOT stop the audio. To account
for that we call parameter stopAudio directly and invert the parameters
in calls.

This allows to keep (for now) LeAudioService's method removeActiveDevice
as is.

Bug: 329809288
Flag: EXEMPT PURE_REFACTOR
Test: atest ActiveDeviceManagerTest
Change-Id: I25225f3758c1be089e670a0bcbdd6494af774d99