Do not report collision for null port
This function is only supposed to report collision if a mux or port is
in progress, and was erroneously reporting a collision with a connected
mux that had no ports associated with it.

This fix changes only one edge case and has a low blast radius, very low
risk of regression.  Additionally, the associated unit test has been
modified for this edge case.  Because of this, not flagging.

Bug: 455435375
Test: atest StackRfcommTest
Flag: EXEMPT BUGFIX
Change-Id: I29ababacda7d8ba0d3d6ed3f3b21ceb7b96fd8f5