Add null checks in l2c_csm
Fuzzer finds two cases where a CCB with a null resource control block can be passed into event handlers in l2c_csm, leading to a null dereference when the RCB is logged.

Both of these bugs relate to the OPEN state, but it should be safe to add null checks in all event handlers where the RCB is used.

Bug: 454422761
Bug: 454115874
Test: m com.google.android.bt
Flag: EXEMPT BUGFIX
Change-Id: I12b83599af85bab95f46203215044be40005d069