VCP: release class lock when calling adapterService
The scope of the lock is bigger than what needed. We don't need a lock
when checking method parameter, nor when calling to get the list of
bonded devices nor reporting the connection change.

I also wonder if we need a lock at all, since most of the other methods
take a lock on mStateMachine only. Locking `this` doesn't seems what we
want but this CL is just keeping the existing lock, just reducin the
scope of it

This Cl is just a short term bandaid. The correct fix would be to handle
thread issue at the root by moving all actions on the looper and
removing the need for any locks

Bug: 443879976
Test: m .
Flag: EXEMPT PURE_REFACTOR
Change-Id: I6a6a6964fa7f60598158e6b3769673e157348db4