Notify connection_manager of Accept List changes
The stack disconnects the GATT link if the remote device
fails to respond within a certain timeout period (e.g., 30 seconds).

This disconnection process removes the device from the Accept List.
However, the `connection_manager` remains unaware of this change and
still considers the device to be "in the Accept List",
which prevents subsequent reconnection attempts.

This change resolves the issue by notifying the `connection_manager` immediately.
This enables the host to correctly re-initiate connections if requested.

Flag: EXEMPT BUGFIX
Bug: 466406346
Test: atest bluetooth_test_gd_unit
Change-Id: I27b0a34961c5cdd43770a5b9264b6eb26930834e