Use the consteval Uuid constructor instead of FromString
Uuid::FromString is used extensively to build the UUIDs
for GATT profiles in the stack. In practice this means
that the UUIDs are constructed at init() since FromString
is not constexpr.

Using the consteval constructor and making all the UUIDs
constinit one ensures that the UUIDs are initialized in
memory at link time (.ro section)

Bug: 331817295
Test: m com.android.bt
Flag: EXEMPT PURE_REFACTOR
Change-Id: I77c9f7844efc2b0face19be9519eef5282d86063