Correct gatt_cmd_dequeue
The `gatt_cmd_dequeue` function is used to verify if
an incoming ATT response matches the expected pending command.

Previously, if the incoming response did not match the command
in the queue, the stack would ignore the response and
errorneously drop the command from the queue.

In scenarios where requests and responses become out of sync,
this behavior prevented the stack from correctly handling
subsequent responses.

This change fixes the issue by using a "peek" operation instead of
"dequeue" when validating the response.

Additionally, this change resolves a potential resource leak.

Flag: com.android.bluetooth.flags.fix_gatt_cmd_dequeue
Bug: 456606891
Bug: 468220386
Test: atest bluetooth_test_gd_unit
Change-Id: I809fc11e87da21089b956b94bef48c40404fab9d