Remove AVRCP target enabled check in AVRC_MsgReq
Concurrent source and sink devices are not correctly packing the
AVRC header bytes. Target implementations skip this step in
AVRC_MsgReq (is_new_avrcp reassigned to true), as they are assumed
to pack these bytes themselves. When a device is both a target and
controller, we are intercepted and do not enter that part of the
target implementation, leading the header bytes to be missing.
Thus, for concurrent source and sink, we should still pack the AVRC
header bytes within AVRC_MsgReq.

In a previous change, the if clause guarding is_new_avrcp assignment
was mistakenly inverted, leading to the header bytes not being
included for concurrent source and sink. However, due to there
already being a parameter that can determine if we are being called
in a target only scenario, this target enabled check is being removed.

Bug: 342488794
Bug: 392193472
Flag: EXEMPT BUGFIX
Test: m com.android.bt
Test: manual Pair a concurrent source and sink device with an iPhone.
Test: In snoop logs, verify the GetCapabilities(Events Supported)
Test: response sent to the iPhone is correctly formatted.
Change-Id: If362aef7f1441388df03deb8c4b7fac3c2a616b7